pyproject.toml
[project]
name = "agent-talk-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "realtimetts==0.2.0",
    "websockets>=15.0.1",
]
